+++ naive_flow/config/utils.py
@@ -6,10 +6,19 @@ from contextlib import contextmanager
 from io import StringIO
 from typing import Literal
 
+import pydantic_settings
 from dotenv import dotenv_values
-from pydantic import Field, version
+from packaging import version
+from pydantic import Field
+from pydantic import version as pydantic_version
 from pydantic_settings import BaseSettings, DotEnvSettingsSource, sources
-from pydantic_settings.sources import parse_env_vars
+
+if version.parse(pydantic_settings.__version__) >= version.parse("2.9"):
+    from pydantic_settings.sources.utils import \
+        parse_env_vars  # pylint: disable=no-name-in-module,import-error
+else:
+    from pydantic_settings.sources import \
+        parse_env_vars  # pylint: disable=no-name-in-module,import-error
 
 
 def strfconfig(
@@ -42,7 +51,9 @@ def strfconfig(
     Returns:
         str: formatted config.
     """
-    if description is not None and float(version.version_short()) < 2.7:  #pylint: disable=no-member
+    if description is not None and float(
+        pydantic_version.version_short()  #pylint: disable=no-member
+    ) < 2.7:
         warnings.warn(
             "Your version of pydantic is before 2.7, "
             "which does not take docstrings of fields as description."